requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
psycopg2-binary
python-dotenv
python-jose[cryptography]
requests
firebase-admin==6.4.0
pydantic-settings>=2.0.0
pydantic[email]
httpx 
redis[asyncio]
slowapi
python-multipart
itsdangerous
jinja2
aiofiles
cloudinary
python-slugify